+++ include/xtensor/xstridedview.hpp
@@ -6,8 +6,8 @@
 * The full license is in the file LICENSE, distributed with this software. *
 ****************************************************************************/
 
-#ifndef XSTRIDEDVIEW_HPP
-#define XSTRIDEDVIEW_HPP
+#ifndef XTENSOR_STRIDEDVIEW_HPP
+#define XTENSOR_STRIDEDVIEW_HPP
 
 #include <algorithm>
 #include <cstddef>