+++ docs/source/changelog.rst
@@ -15,6 +15,8 @@ Breaking changes
 
 - The API for ``xbuffer_adaptor`` has changed. The template parameter is the type of the buffer, not just the value type.
   `#482 <https://github.com/QuantStack/xtensor/pull/482>`_
+- Change ``edge_items`` print option to ``edgeitems`` for better numpy consistency.
+  `#489 <https://github.com/QuantStack/xtensor/pull/489>`_.
 - xtensor now depends on ``xtl`` version `~0.3.1`.
 
 Other changes